Move FakeService to a "testing" package.
