+++ provider/local/export_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/instance"
 	"github.com/juju/juju/mongo"
-	"github.com/juju/juju/service"
 	"github.com/juju/juju/service/common"
+	svctesting "github.com/juju/juju/service/common/testing"
 )
 
 var (
@@ -89,7 +89,7 @@ func PatchCreateContainer(s *testing.CleanupSuite, c *gc.C, expectedURL string) 
 	return mockFunc
 }
 
-func PatchServices(patchValue func(interface{}, interface{}), data *service.FakeServiceData) {
+func PatchServices(patchValue func(interface{}, interface{}), data *svctesting.FakeServiceData) {
 	patchValue(&mongoRemoveService, func(namespace string) error {
 		data.AddCall("RemoveService", namespace)
 		data.SetStatus(mongo.ServiceName(namespace), "")
@@ -100,8 +100,8 @@ func PatchServices(patchValue func(interface{}, interface{}), data *service.Fake
 	})
 }
 
-func NewService(name string, conf common.Conf, data *service.FakeServiceData) *service.FakeService {
-	svc := service.NewFakeService(name, conf)
+func NewService(name string, conf common.Conf, data *svctesting.FakeServiceData) *svctesting.FakeService {
+	svc := svctesting.NewFakeService(name, conf)
 	svc.FakeServiceData = data
 	return svc
 }
