+++ service/common/testing/fake.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package service
+package testing
 
 import (
 	"strings"
@@ -86,7 +86,7 @@ func (fsd *FakeServiceData) SetStatus(name, status string) error {
 	return nil
 }
 
-var _ Service = (*FakeService)(nil)
+//var _ Service = (*FakeService)(nil)
 
 // FakeService is a Service implementation for testing.
 type FakeService struct {
