+++ service/common/testing/interface_test.go
@@ -1,0 +1,11 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the LGPLv3, see LICENCE file for details.
+
+package testing_test
+
+import (
+	"github.com/juju/juju/service"
+	"github.com/juju/juju/service/common/testing"
+)
+
+var _ service.Service = (*testing.FakeService)(nil)
