+++ service/interface_test.go
@@ -1,3 +1,6 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the LGPLv3, see LICENCE file for details.
+
 package service
 
 import (
