+++ service/service_test.go
@@ -11,6 +11,7 @@ import (
 	gc "gopkg.in/check.v1"
 
 	"github.com/juju/juju/service"
+	svctesting "github.com/juju/juju/service/common/testing"
 	"github.com/juju/juju/service/systemd"
 	"github.com/juju/juju/service/upstart"
 	"github.com/juju/juju/service/windows"
@@ -136,7 +137,7 @@ func (s *restartSuite) TestRestartStopAndStart(c *gc.C) {
 }
 
 type restartable struct {
-	*service.FakeService
+	*svctesting.FakeService
 }
 
 func (s *restartable) Restart() error {
