+++ worker/deployer/export_test.go
@@ -6,8 +6,8 @@ package deployer
 import (
 	"github.com/juju/juju/agent"
 	"github.com/juju/juju/apiserver/params"
-	"github.com/juju/juju/service"
 	"github.com/juju/juju/service/common"
+	svctesting "github.com/juju/juju/service/common/testing"
 )
 
 type fakeAPI struct{}
@@ -19,12 +19,12 @@ func (*fakeAPI) ConnectionInfo() (params.DeployerConnectionValues, error) {
 	}, nil
 }
 
-func NewTestSimpleContext(agentConfig agent.Config, logDir string, data *service.FakeServiceData) *SimpleContext {
+func NewTestSimpleContext(agentConfig agent.Config, logDir string, data *svctesting.FakeServiceData) *SimpleContext {
 	return &SimpleContext{
 		api:         &fakeAPI{},
 		agentConfig: agentConfig,
 		discoverService: func(name string, conf common.Conf) deployerService {
-			svc := service.NewFakeService(name, conf)
+			svc := svctesting.NewFakeService(name, conf)
 			svc.FakeServiceData = data
 			return svc
 		},
